@@ -1,13 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './styles/index.css';
-import App from './components/App';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
+import WrappedApp from './components/App';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 //import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
+  
   <Router>
-    <Route path="/" component={App} />
+    <Routes>
+      <Route path="/" element={<WrappedApp />} />
+    </Routes>
   </Router>,
   document.getElementById('root')
 );
